식당 등록 + 지역 등록 서비스 구현 완료(더미데이터 생성 완료)

+++ kh16fd/src/main/java/com/kh/fd/dto/PlaceDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class PlaceDto {
-    private int placeId;
+    private Long placeId;
     private String placeDepth1;
     private String placeDepth2;
     private String placeDepth3;
